+++ package.json
@@ -51,7 +51,7 @@
     "streaming-iterables": "^4.1.2"
   },
   "peerDependencies": {
-    "libp2p": "libp2p/js-libp2p#feat/certified-addressbook"
+    "libp2p": "libp2p/js-libp2p#0.29.x"
   },
   "devDependencies": {
     "aegir": "^25.0.0",
@@ -59,7 +59,7 @@
     "chai-as-promised": "^7.1.1",
     "delay": "^4.3.0",
     "dirty-chai": "^2.0.1",
-    "libp2p": "libp2p/js-libp2p#feat/certified-addressbook",
+    "libp2p": "libp2p/js-libp2p#0.29.x",
     "libp2p-mplex": "^0.9.5",
     "libp2p-noise": "^1.1.2",
     "libp2p-websockets": "^0.13.6",
