+++ test/rendezvous.spec.js
@@ -6,7 +6,9 @@ chai.use(require('dirty-chai'))
 chai.use(require('chai-as-promised'))
 const { expect } = chai
 const sinon = require('sinon')
+const pWaitFor = require('p-wait-for')
 
+const multiaddr = require('multiaddr')
 const Envelope = require('libp2p/src/record/envelope')
 const PeerRecord = require('libp2p/src/record/peer-record')
 
@@ -14,6 +16,8 @@ const Rendezvous = require('../src')
 const { codes: errCodes } = require('../src/errors')
 
 const { createPeer, connectPeers } = require('./utils')
+const { MULTIADDRS_WEBSOCKETS } = require('./fixtures/browser')
+const relayAddr = MULTIADDRS_WEBSOCKETS[0]
 
 const namespace = 'ns'
 
@@ -240,6 +244,95 @@ describe('rendezvous', () => {
         registers.push(reg)
       }
 
+      expect(registers).to.have.lengthOf(1)
+    })
+  })
+
+  describe('flows with 3 rendezvous all acting as rendezvous point', () => {
+    let peers
+
+    const connectPeers = async (peer, otherPeer) => {
+      // Connect each other via relay node
+      const m = multiaddr(`${relayAddr}/p2p-circuit/p2p/${otherPeer.peerId.toB58String()}`)
+      await peer.dial(m)
+
+      // Wait event propagation
+      await pWaitFor(() => peer.rendezvous._rendezvousPoints.size === 1)
+    }
+
+    beforeEach(async () => {
+      // Create libp2p nodes
+      peers = await createPeer({
+        number: 3
+      })
+
+      // Create 3 rendezvous peers
+      peers.forEach((peer) => {
+        const rendezvous = new Rendezvous({
+          libp2p: peer
+        })
+        rendezvous.start()
+        peer.rendezvous = rendezvous
+      })
+
+      // Connect to testing relay node
+      await Promise.all(peers.map((libp2p) => libp2p.dial(relayAddr)))
+    })
+
+    afterEach(() => peers.map(async (libp2p) => {
+      await libp2p.rendezvous.stop()
+      await libp2p.stop()
+    }))
+
+    it('discover find registered peer for namespace only when registered', async () => {
+      await connectPeers(peers[0], peers[1])
+      await connectPeers(peers[2], peers[1])
+
+      const registers = []
+
+      // Peer2 does not discovery any peer registered
+      for await (const reg of peers[2].rendezvous.discover(namespace)) { // eslint-disable-line
+        throw new Error('no registers should exist')
+      }
+
+      // Peer0 register itself on namespace (connected to Peer1)
+      await peers[0].rendezvous.register(namespace)
+
+      // Peer2 discovers Peer0 registered in Peer1
+      for await (const reg of peers[2].rendezvous.discover(namespace)) {
+        registers.push(reg)
+      }
+      expect(registers).to.have.lengthOf(1)
+      expect(registers[0].signedPeerRecord).to.exist()
+      expect(registers[0].ns).to.eql(namespace)
+      expect(registers[0].ttl).to.exist()
+
+      // Peer0 unregister itself on namespace (connected to Peer1)
+      await peers[0].rendezvous.unregister(namespace)
+
+      // Peer2 does not discovery any peer registered
+      for await (const reg of peers[2].rendezvous.discover(namespace)) { // eslint-disable-line
+        throw new Error('no registers should exist')
+      }
+    })
+
+    it('discovers locally first, and if limit achieved, not go to the network', async () => {
+      await connectPeers(peers[0], peers[1])
+      await connectPeers(peers[2], peers[1])
+
+      // Peer0 register itself on namespace (connected to Peer1)
+      await peers[1].rendezvous.register(namespace)
+
+      const spyRendezvousPoints = sinon.spy(peers[2].rendezvous._rendezvousPoints, 'entries')
+
+      const registers = []
+      // Peer2 discovers Peer0 registered in Peer1
+      for await (const reg of peers[2].rendezvous.discover(namespace, 1)) {
+        registers.push(reg)
+      }
+
+      // No need to get the rendezvousPoints connections
+      expect(spyRendezvousPoints).to.have.property('callCount', 0)
       expect(registers).to.have.lengthOf(1)
     })
   })
